+++ src/ploomber_engine/profiling.py
@@ -115,3 +115,28 @@ def plot_cell_runtime(nb):
     ax.figure.tight_layout()
 
     return ax
+
+
+def get_profiling_data(nb):
+    """
+    Collect profiling data from the notebook and return as a dictionary
+
+    Parameters
+    ----------
+    nb : NotebookNode
+        Notebook to collect profiling data from
+
+    Returns
+    -------
+    dict: dictionary with the following keys:
+        cell: list of cell indexes
+        runtime: list of cell runtimes
+        memory: list of cell memory usage
+    """
+    code_cells = [cell for cell in nb.cells if cell.cell_type == "code"]
+    _get_mem = lambda c: c.metadata["ploomber"].get("memory_usage", "NA")
+    return dict(
+        cell=list(range(1, len(code_cells) + 1)),
+        runtime=[_compute_runtime(c) for c in code_cells],
+        memory=[_get_mem(c) for c in code_cells],
+    )